+++ wizeline-os-ai-tools/src/components/Form/ContactForm.tsx
@@ -1,132 +1,134 @@
-import { Input } from "../ui/Input";
+import { FC } from 'react'
+import { Controller, useFormContext } from "react-hook-form";
+import { TextField,} from "@material-ui/core";
 import { Heading, headingVariants } from "../ui/Heading";
 import { cn } from "@/utils/utils";
+import Label from '../ui/Label';
 
-type Props = {
-  formikProps: {
-    values: {
-      fullName: string;
-      title: string;
-      country: string;
-      state: string;
-      city: string;
-      phoneNumber: number;
-      avatarURL: string;
-    };
-  };
-};
+interface ContactForm2Props {
+  
+}
 
-const ContactForm = ({ handleChange, values, errors, touched }) => {
+const ContactForm2: FC<ContactForm2Props> = ({}) => {
+    const { control, formState: { errors }, watch } = useFormContext();
+
   return (
     <div className="flex flex-col gap-4">
-      <Heading className={cn(headingVariants({ size: "default" }))}>
-        {" "}
-        Contact{" "}
-      </Heading>
-		<Input
-			type="text"
-			title="Full Name*"
-			name="fullName"
-			value={values.fullName}
-			onChange={handleChange}
-			className={` ${
-			errors.fullName && touched.fullName ? "border-2 border-rose-600" : ""
-			}`}
-		/>
-		{errors.fullName && touched.fullName && (
-			<p className="text-sm text-pink-600">{errors.fullName}</p>
-		)}
-		<Input
-			type="text"
-			title="Title"
-			name="title"
-			value={values.title}
-			onChange={handleChange}
-			className={` ${
-			errors.title && touched.title ? "border-2 border-rose-600" : ""
-			}`}
-		/>
-		{errors.title && touched.title && (
-			<p className="text-sm text-pink-600">{errors.title}</p>
-		)}
-		<div className="grid grid-cols-3 gap-4">
-			<div>
-			<Input
-				type="text"
-				title="Country"
-				name="country"
-				value={values.country}
-				onChange={handleChange}
-				className={` ${
-				errors.country && touched.country
-					? "border-2 border-rose-600"
-					: ""
-				}`}
-			/>
-			{errors.country && touched.country && (
-				<p className="text-sm text-pink-600">{errors.country}</p>
-			)}
-			</div>
-			<div>
-			<Input
-				type="text"
-				title="State"
-				name="state"
-				value={values.state}
-				onChange={handleChange}
-				className={` ${
-				errors.state && touched.state ? "border-2 border-rose-600" : ""
-				}`}
-			/>
-			{errors.state && touched.state && (
-				<p className="text-sm text-pink-600">{errors.state}</p>
-			)}
-			</div>
-			<div>
-			<Input
-				type="text"
-				title="City"
-				name="city"
-				value={values.city}
-				onChange={handleChange}
-				className={` ${
-				errors.city && touched.city ? "border-2 border-rose-600" : ""
-				}`}
-			/>
-			{errors.city && touched.city && (
-				<p className="text-sm text-pink-600">{errors.city}</p>
-			)}
-			</div>
-		</div>
-		<Input
-			type="text"
-			title="Phone *"
-			name="phoneNumber"
-			value={values.phoneNumber}
-			onChange={handleChange}
-			className={` ${
-			errors.phoneNumber && touched.phoneNumber ? "border-2 border-rose-600" : ""
-			}`}
-		/>
-		{errors.phoneNumber && touched.phoneNumber && (
-			<p className="text-sm text-pink-600">{errors.phoneNumber}</p>
-		)}
-		<Input
-			type="text"
-			title="Avatar URL"
-			name="avatarURL"
-			value={values.avatarURL}
-			onChange={handleChange}
-			className={` ${
-			errors.avatarURL && touched.avatarURL ? "border-2 border-rose-600" : ""
-			}`}
-		/>
-		{errors.avatarURL && touched.avatarURL && (
-			<p className="text-sm text-pink-600">{errors.avatarURL}</p>
-		)}
+        <Heading className={cn(headingVariants({ size: "default" }))}> Contact </Heading>
+        <div>
+            <Label title='Full Name'></Label>
+            <Controller
+                control={control}
+                name="fullName"
+                rules={{ required: "This field is required." }}
+                render={({ field }) => (
+                    <TextField
+                    id="fullName"
+                    variant="outlined"
+                    fullWidth
+                    {...field}
+                    error={Boolean(errors?.fullName)}
+                    helperText={errors.fullName?.message?.toString() || '' }
+                    />
+                )}
+            />
+        </div>
+        <div className="grid grid-cols-3 gap-4">
+            <div>
+                <Label title='Country'></Label>
+                <Controller
+                    control={control}
+                    name="country"
+                    rules={{ required: "This field is required." }}
+                    render={({ field }) => (
+                        <TextField
+                        id="country"
+                        variant="outlined"
+                        fullWidth
+                        {...field}
+                        error={Boolean(errors?.country)}
+                        helperText={errors.country?.message?.toString() || '' }
+                        />
+                    )}
+                />
+            </div>
+            <div>
+                <Label title='State'></Label>
+                <Controller
+                    control={control}
+                    name="state"
+                    rules={{ required: "This field is required." }}
+                    render={({ field }) => (
+                        <TextField
+                        id="state"
+                        variant="outlined"
+                        fullWidth
+                        {...field}
+                        error={Boolean(errors?.state)}
+                        helperText={errors.state?.message?.toString() || '' }
+                        />
+                    )}
+                />
+                {/* {errors.state && touched.state && ( <p className="text-sm text-pink-600">{errors.state}</p> )} */}
+            </div>
+            <div>
+                <Label title='City'></Label>
+                <Controller
+                    control={control}
+                    name="city"
+                    rules={{ required: "This field is required." }}
+                    render={({ field }) => (
+                        <TextField
+                        id="city"
+                        variant="outlined"
+                        fullWidth
+                        {...field}
+                        error={Boolean(errors?.city)}
+                        helperText={errors.city?.message?.toString() || '' }
+                        />
+                    )}
+                />
+            </div>
+        </div>
+        <div>
+            <Label title='Phone Number'></Label>
+            <Controller
+                control={control}
+                name="phoneNumber"
+                rules={{ required: "This field is required." }}
+                render={({ field }) => (
+                    <TextField
+                    id="phoneNumber"
+                    variant="outlined"
+                    fullWidth
+                    {...field}
+                    error={Boolean(errors?.phoneNumber)}
+                    helperText={errors.phoneNumber?.message?.toString() || '' }
+                    />
+                )}
+            />
+        </div>
+        <div>
+            <Label title='Avatar URL'></Label>
+            <Controller
+                control={control}
+                name="avatarURL"
+                rules={{ required: "This field is required." }}
+                render={({ field }) => (
+                    <TextField
+                    id="avatarURL"
+                    variant="outlined"
+                    fullWidth
+                    {...field}
+                    error={Boolean(errors?.avatarURL)}
+                    helperText={errors.avatarURL?.message?.toString() || '' }
+                    />
+                )}
+            />
+        </div>
 
-    </div>
-  );
-};
+  </div>
+  )
+}
 
-export default ContactForm;
+export default ContactForm2
