+++ wizeline-os-ai-tools/src/components/Form/FormComponent.tsx
@@ -1,121 +1,106 @@
-import { Formik, Form } from "formik";
+import { FC } from 'react'
+import { useForm, FormProvider } from "react-hook-form";
 import * as Yup from "yup";
-import EducationForm from "../Form/EducationForm";
-import ContactForm from "./ContactForm";
-import PastWorkForm from "./PastWorkForm";
-import SkillsForm from "./SkillsForm";
-import AboutForm from "./AboutForm";
-import { Input } from "../ui/Input";
-import { Button, buttonVariants } from "../ui/Button";
-import { Heading } from "../ui/Heading";
-import AIAssitant from "../AIAssistant";
+import { yupResolver } from '@hookform/resolvers/yup';
+
+import { Heading } from '../ui/Heading';
+import { Button, buttonVariants } from '../ui/Button';
+import AboutForm from './AboutForm';
+import ContactForm from './ContactForm';
+import PastWorkForm from './PastWorkForm';
+import EducationForm from './EducationForm';
+import SkillsForm from './SkillsForm';
+import { SkillsOptions } from '@/utils/skillsData';
+
+interface FormValues{
+    aiAssistant: any,
+    aboutDescription: string,
+    fullName: string,
+    title: string,
+    country: string,
+    state: string,
+    city: string,
+    phoneNumber:number,
+    avatarURL: string,
+    pastWtitle: string,
+    pastWDescription: string,
+    pastWStartDate: string,
+    pastWEndDate: string,
+    schoolName: string,
+    degree: string,
+    specialization1: string,
+    specialization2: string,
+    expertSkills: SkillsOptions[],
+    advancedSkills: SkillsOptions[],
+    intermediateSkills: SkillsOptions[],
+    basicSkills: SkillsOptions[],
+}
 
 const validationSchema = Yup.object().shape({
-  aboutDescription: Yup.string().required("Description is required").max(500),
-  fullName: Yup.string().required("Full Name is required"),
-  title: Yup.string().required("Title is required"),
-  country: Yup.string().required("Country is required"),
-  state: Yup.string().required("State is required"),
-  city: Yup.string().required("City is required"),
-  phoneNumber: Yup.number().required("Phone is required"),
-  avatarURL: Yup.string().required("Avatar URL is required"),
-  schoolName: Yup.string().required("School Name is required"),
-  degree: Yup.string().required("Degree is required"),
+    aboutDescription: Yup.string().required("Description is required").max(800),
+    fullName: Yup.string().required("Full Name is required"),
+    title: Yup.string().required("Title is required"),
+    country: Yup.string().required("Country is required"),
+    state: Yup.string().required("State is required"),
+    city: Yup.string().required("City is required"),
+    phoneNumber: Yup.number().required("Phone is required"),
+    avatarURL: Yup.string().required("Avatar URL is required"),
+    pastWtitle: Yup.string().required("Title is required"),
+    pastWDescription: Yup.string().required("Description is required").max(800),
+    pastWStartDate: Yup.string().required("Start Date is required"),
+    pastWEndDate: Yup.string().required("End Date is required"),
+    schoolName: Yup.string().required("School Name is required"),
+    degree: Yup.string().required("Degree is required"),
 });
 
-const FormComponent = () => {
+interface FormComponent2Props {
+    
+}
 
-  return (
-    <Formik
-      initialValues={{
-        aboutDescription: "",
-        fullName: "",
-        title: "",
-        country: "",
-        state: "",
-        city: "",
-        phoneNumber: 0,
-        avatarURL: "",
-        schoolName: "",
-        degree: "",
-        specialization1: "",
-        specialization2: "",
-        pastWDescription: "",
-      }}
-      validationSchema={validationSchema}
-      
-      onSubmit={(values, actions) => {
-        console.log(values);
-      }}
-    >
-      {({ handleSubmit, handleChange, values, errors, touched }) => (
-        <Form className="container mx-auto">
-          <div className="grid grid-cols-9">
-            <div className="-bg-orange-500 col-span-6 m-8 flex flex-col gap-8">
-              <Heading>Hello, name!</Heading>
-              <p className="text-base font-light text-gray-400">
-                Fill out the following information to create your profile. If
-                you changed your mind and want to create you profile with
-                Linkedin, make sure to hit the button and start with the
-                process.
-              </p>
-              <div className="w-52">
-                <Button
-                  className={buttonVariants({
-                    variant: "linkedin",
-                    size: "logIn",
-                  })}
-                >
-                  <p className="">Create with Linkedin</p>
-                </Button>
-              </div>
+const FormComponent2: FC<FormComponent2Props> = ({}) => {
+    const methods = useForm<FormValues>({
+        //resolver: yupResolver(validationSchema),
+    })
 
-              <AboutForm
-                handleChange={handleChange}
-                values={values}
-                errors={errors}
-                touched={touched}
-              ></AboutForm>
+    const onSubmit = (data: FormValues) => {
+        console.log(data);
+    };
 
-              <ContactForm
-                handleChange={handleChange}
-                values={values}
-                errors={errors}
-                touched={touched}
-              ></ContactForm>
+    return (
+        <FormProvider {...methods}>
+            <form onSubmit={methods.handleSubmit(onSubmit)} className="container mx-auto" >
+                <div className="grid grid-cols-9">
+                    <div className="-bg-orange-500 col-span-6 m-8 flex flex-col gap-8">
+                        <Heading>Hello, name!</Heading>
+                        <p className="text-base font-light text-gray-400">
+                            Fill out the following information to create your profile. If
+                            you changed your mind and want to create you profile with
+                            Linkedin, make sure to hit the button and start with the
+                            process.
+                        </p>
+                        <div className="w-52">
+                            <Button
+                            className={buttonVariants({
+                                variant: "linkedin",
+                                size: "logIn",
+                            })}
+                            >
+                            <p className="">Create with Linkedin</p>
+                            </Button>
+                        </div>
 
-              <PastWorkForm
-                handleChange={handleChange}
-                values={values}
-                errors={errors}
-                touched={touched}
-              ></PastWorkForm>
+                        <AboutForm></AboutForm>
+                        <ContactForm></ContactForm>
+                        <PastWorkForm></PastWorkForm>
+                        <EducationForm></EducationForm>
+                        <SkillsForm></SkillsForm>
 
-              <EducationForm
-                handleChange={handleChange}
-                values={values}
-                errors={errors}
-                touched={touched}
-              ></EducationForm>
-
-              <SkillsForm></SkillsForm>
-
-              <div className="grid justify-items-end">
-                <div className="w-52">
-                  <Input type="submit" value="Submit"></Input>
+                        <Button type="submit" className={buttonVariants({variant: 'linkedin', size: 'logIn'})}></Button>
+                    </div>
                 </div>
-              </div>
-            </div>
-            <div className="col-span-3 bg-gray-200 pl-7 pr-5 pt-64">
-              <AIAssitant
-                aboutText={values.aboutDescription}
-              />
-            </div>
-          </div>
-        </Form>
-      )}
-    </Formik>
-  );
-};
+            </form>
+        </FormProvider>
+    )
+}
 
-export default FormComponent;
+export default FormComponent2
