+++ wizeline-os-ai-tools/src/components/Form/PastWorkForm.tsx
@@ -1,99 +1,115 @@
-import { Input } from "../ui/Input";
+import { FC } from 'react'
+import { Controller, useFormContext } from "react-hook-form";
+import { TextField,} from "@material-ui/core";
 import { Heading, headingVariants } from "../ui/Heading";
 import { cn } from "@/utils/utils";
+import Label from '../ui/Label';
 
-type Props = {
-  formikProps: {
-    values: {
-      pastWTitle: string;
-      startDate: string;
-      endDate: string;
-      pastWDescription: string;
-    };
-  };
-};
+interface PastWorkForm2Props {
+  
+}
 
-const PastWorkForm = ({ handleChange, values, errors, touched }) => {
-  return (
-    <div className="flex flex-col gap-4">
-      <Heading className={cn(headingVariants({ size: "default" }))}>
-        {" "}
-        Past Work{" "}
-      </Heading>
-      <Input
-        type="text"
-        title="Title *"
-        name="pastWtitle"
-        value={values.pastWtitle}
-        onChange={handleChange}
-        className={` ${
-          errors.pastWtitle && touched.pastWtitle
-            ? "border-2 border-rose-600"
-            : ""
-        }`}
-      />
-      {errors.pastWtitle && touched.pastWtitle && (
-        <p className="text-sm text-pink-600">{errors.pastWtitle}</p>
-      )}
-      <div className="grid grid-cols-2 gap-4">
-        <div>
-          <Input
-            type="text"
-            title="Start Date"
-            name="startDate"
-            value={values.startDate}
-            onChange={handleChange}
-            className={` ${
-              errors.startDate && touched.startDate
-                ? "border-2 border-rose-600"
-                : ""
-            }`}
-          />
-          {errors.startDate && touched.startDate && (
-            <p className="text-sm text-pink-600">{errors.startDate}</p>
-          )}
+const PastWorkForm2: FC<PastWorkForm2Props> = ({}) => {
+    const { control, formState: { errors }, watch } = useFormContext();
+
+    const pastWDescription = watch("pastWDescription")
+    const characterLimit = 800;
+    const isExceededLimit = pastWDescription && pastWDescription.length > characterLimit;
+
+    return (
+        <div className="flex flex-col gap-4">
+            <Heading className={cn(headingVariants({ size: "default" }))}>
+                {" "}
+                Past Work{" "}
+            </Heading>
+            <div>
+                <Label title='Title'></Label>
+                <Controller
+                    control={control}
+                    name="pastWtitle"
+                    rules={{ required: "This field is required." }}
+                    render={({ field }) => (
+                        <TextField
+                        id="pastWtitle"
+                        variant="outlined"
+                        fullWidth
+                        {...field}
+                        error={Boolean(errors?.pastWtitle)}
+                        helperText={errors.pastWtitle?.message?.toString() || '' }
+                        />
+                    )}
+                />
+                {/* {errors.pastWtitle && touched.pastWtitle && ( <p className="text-sm text-pink-600">{errors?.pastWtitle.toString()}</p>)} */}
+            </div>
+            <div className="grid grid-cols-2 gap-4">
+                <div>
+                    <Label title='Start Date'></Label>
+                    <Controller
+                        control={control}
+                        name="pastWStartDate"
+                        rules={{ required: "This field is required." }}
+                        render={({ field }) => (
+                            <TextField
+                            id="pastWStartDate"
+                            variant="outlined"
+                            fullWidth
+                            {...field}
+                            error={Boolean(errors?.pastWStartDate)}
+                            helperText={errors.pastWStartDate?.message?.toString() || '' }
+                            />
+                        )}
+                    />
+                    {/* {errors.pastWStartDate && touched.pastWStartDate && ( <p className="text-sm text-pink-600">{errors.pastWStartDate}</p>)} */}
+                </div>
+                <div>
+                    <Label title='End Date'></Label>
+                    <Controller
+                        control={control}
+                        name="pastWEndDate"
+                        rules={{ required: "This field is required." }}
+                        render={({ field }) => (
+                            <TextField
+                            id="pastWEndDate"
+                            variant="outlined"
+                            fullWidth
+                            {...field}
+                            error={Boolean(errors?.pastWEndDate)}
+                            helperText={errors.pastWEndDate?.message?.toString() || '' }
+                            />
+                        )}
+                    />
+                    {/* {errors.endDate && touched.endDate && ( <p className="text-sm text-pink-600">{errors.endDate}</p>)} */}
+                </div>
+            </div>
+            <div>
+                <Label title='Description'></Label>
+                <div
+                    className={` ${
+                    errors.pastWDescription || isExceededLimit
+                        ? "block w-full rounded border border-rose-600 p-3.5 text-sm text-gray-900 focus:border-rose-600 focus:ring-rose-600"
+                        : "block w-full rounded border border-gray-300 p-3.5 text-sm text-gray-900 focus:border-blue-500 focus:ring-blue-500"
+                    }`}
+                >
+                    <Controller
+                        control={control}
+                        name="pastWDescription"
+                        rules={{ required: "This field is required." }}
+                        render={({ field }) => (
+                            <textarea
+                            rows={5}
+                            id="pastWDescription"
+                            className="w-full resize-none"
+                            {...field}
+                            />
+                        )}
+                    />
+                    <p className= {` ${ errors.pastWDescription || isExceededLimit ? "text-right text-rose-600" : "text-right text-gray-400" }`}>
+                        { watch('pastWDescription') ? (watch('pastWDescription').length) : "0" }/{characterLimit}
+                    </p>
+                </div>
+            </div>
         </div>
-        <div>
-          <Input
-            type="text"
-            title="End Date"
-            name="endDate"
-            value={values.endDate}
-            onChange={handleChange}
-            className={` ${
-              errors.endDate && touched.endDate
-                ? "border-2 border-rose-600"
-                : ""
-            }`}
-          />
-          {errors.endDate && touched.endDate && (
-            <p className="text-sm text-pink-600">{errors.endDate}</p>
-          )}
-        </div>
-      </div>
-      <div>
-        <label
-          htmlFor=""
-          className="text-md mb-1 block font-medium text-gray-900"
-        >
-        Description
-        </label>
-		<div className='border border-gray-300 text-gray-900 text-sm rounded focus:ring-blue-500 focus:border-blue-500 block w-full p-3.5'>
-		<textarea
-          name="pastWDescription"
-          placeholder="Description..."
-          value={values.pastWDescription}
-          onChange={handleChange}
-          className='resize-none w-full'
-          rows={5}
-        />
-		<p className="text-right text-gray-400">
-          {values.pastWDescription.length}/500
-        </p>
-		</div>
-      </div>
-    </div>
-  );
-};
+    );
+}
 
-export default PastWorkForm;
+export default PastWorkForm2
