+++ wizeline-os-ai-tools/src/pages/newUser.tsx
@@ -1,5 +1,4 @@
 import FormComponent from "@/components/Form/FormComponent";
-import FormComponent2 from "@/components/Form/FormComponent2";
 export default function newUser() {
     return(
         <div className="w-full flex items-start flex-col">
@@ -12,8 +11,7 @@ export default function newUser() {
             <div className="bg-black w-full h-20">
                 pseudoNavbar
             </div>
-            {/* <FormComponent></FormComponent> */}
-            <FormComponent2></FormComponent2>
+            <FormComponent></FormComponent>
         </div>
     )
 }
